requirements.txt
# local package
-e .

# external requirements
mailbox
beautifulsoup4
duckdb
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
